requirements.txt
babel==2.9.0
python-dateutil==2.6.0
flask-moment==0.11.0
flask-wtf==0.14.3
flask_sqlalchemy==2.4.4
absl-py==0.12.0
alembic==1.6.5
appdirs==1.4.4
astroid==2.4.2
astunparse==1.6.3
autopep8==1.5.4
Babel==2.9.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
itsdangerous==2.0.1
Jinja2==3.0.1
jwt==1.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.0.25
oauthlib==3.1.0
opt-einsum==3.3.0
postgres==3.0.0
protobuf==3.15.8
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
python-dateutil==2.6.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.20
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc1
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.7
Werkzeug==2.0.0
wrapt==1.12.1
WTForms==2.3.3
